initiate metrics module and update cli module
