+++ torecsys/metrics/__init__.py
@@ -1,3 +1,5 @@
 r"""torecsys.metrics is a sub module to implemente functions to calculate metrcis of 
 evaluateion in recommendation system.
 """
+
+from .metrics import *
